# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "UltraFit 多租户权限管理系统 API",
        "title": "UltraFit API",
        "termsOfService": "https://ultrafit.com/terms/",
        "contact": {
            "name": "UltraFit Team",
            "email": "admin@ultrafit.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/field-permissions/roles/{roleId}/{tableName}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定角色在指定表的字段权限配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "field-permissions"
                ],
                "summary": "获取角色字段权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "tableName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批量更新角色在指定表的字段权限配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "field-permissions"
                ],
                "summary": "更新角色字段权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "tableName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "字段权限配置",
                        "name": "permissions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_internal_dto.UpdateRoleFieldPermissionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/field-permissions/tables/{tableName}/fields": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定表的字段权限配置列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "field-permissions"
                ],
                "summary": "获取表字段配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "tableName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/system/field-permissions/{tableName}/initialize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定表初始化字段权限配置，需要系统管理员权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "初始化表字段权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "tableName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "字段配置列表",
                        "name": "fields",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_internal_dto.InitializeFieldsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/field-permissions/{tableName}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户对指定表的字段权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取当前用户字段权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "tableName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_varluffy_shield_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_varluffy_shield_internal_dto.FieldConfig": {
            "type": "object",
            "required": [
                "field_name",
                "field_type"
            ],
            "properties": {
                "default_value": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "field_label": {
                    "type": "string"
                },
                "field_name": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "is_required": {
                    "type": "boolean"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "github_com_varluffy_shield_internal_dto.FieldPermissionConfig": {
            "type": "object",
            "required": [
                "field_name",
                "permission_type"
            ],
            "properties": {
                "field_name": {
                    "type": "string"
                },
                "permission_type": {
                    "type": "string"
                }
            }
        },
        "github_com_varluffy_shield_internal_dto.InitializeFieldsRequest": {
            "type": "object",
            "required": [
                "fields"
            ],
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_varluffy_shield_internal_dto.FieldConfig"
                    }
                }
            }
        },
...
        "github_com_varluffy_shield_internal_dto.UpdateRoleFieldPermissionsRequest": {
            "type": "object",
            "required": [
                "field_permissions"
            ],
            "properties": {
                "field_permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_varluffy_shield_internal_dto.FieldPermissionConfig"
                    }
                }
            }
        },
...
        "github_com_varluffy_shield_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "响应时间戳 (RFC3339格式)",
                    "type": "string"
                },
                "trace_id": {
                    "description": "追踪ID",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}